Change name of Notification component
